enchantments on profileManagement work.
