+++ data/profiles.json
@@ -1,6 +1,6 @@
 [{
-    "Bio": "ssss",
+    "Bio": "i am ahmed sameh.",
     "userId": "U1",
-    "ProfilePicture": "images/I1.png",
-    "CoverPhoto": "images/I2.png"
+    "ProfilePicture": "images/I3.png",
+    "CoverPhoto": "images/I4.jpeg"
 }]
