+++ data/users.json
@@ -1,5 +1,5 @@
 [{
-    "password": "wxYOOAE0XNe1JeHMFf1MKsUCaXeKY5BnXfu+qg0kArQ=",
+    "password": "TaEnI1DzlKqvOxSNAXx4ujTvDU5eW9thCt21sUvFa94=",
     "dateOfBirth": "2004-10-12",
     "userId": "U1",
     "email": "a.samehmepsn@ssgfmx.com",
