--- src/main/java/backend/IAddProfile.java
@@ -1,5 +1,0 @@
-package backend;
-
-public interface IAddProfile {
-    public void addProfile(String username, String password, String email, String dob);
-}
