+++ src/main/java/backend/ProfileManager.java
@@ -7,28 +7,17 @@ import java.io.IOException;
 
 
 public class ProfileManager {
-    IAddProfile ProfileAdder;
     ILoadProfiles ProfileLoader;
-    Validation validation;
     String userId;
     IUpdateProfile ProfileUpdater;
 
-    public ProfileManager(IAddProfile addProfile, ILoadProfiles loadProfiles, String userId, Validation validation, IUpdateProfile updateProfile) {
-        this.ProfileAdder = addProfile;
+    public ProfileManager(ILoadProfiles loadProfiles, String userId, IUpdateProfile updateProfile) {
         this.ProfileLoader = loadProfiles;
-        this.validation = validation;
         this.userId = userId;
         this.ProfileUpdater = updateProfile;
     }
-
-    public void add(String ProfilePicture, String CoverPhoto, String Bio) throws IOException {
-        ProfileAdder.addProfile(ProfilePicture, CoverPhoto, Bio, userId);
-        updateProfilePhoto(ProfilePicture);
-        updateCoverPhoto(CoverPhoto);
-    }
-
     public void updateProfilePhoto(String absoluteFilePath) throws IOException {
-        JSONObject profile = validation.findProfileByUserId(userId);
+        JSONObject profile = findProfileByUserId(userId);
         JSONArray profiles = ProfileLoader.loadProfiles();
         String savedImagePath = SaveImage.saveImageToFolder(absoluteFilePath);
         profile.put("ProfilePicture", savedImagePath);
@@ -37,7 +26,7 @@ public class ProfileManager {
     }
 
     public void updateCoverPhoto(String absoluteFilePath) throws IOException {
-        JSONObject profile = validation.findProfileByUserId(userId);
+        JSONObject profile = findProfileByUserId(userId);
         JSONArray profiles = ProfileLoader.loadProfiles();
         String savedImagePath = SaveImage.saveImageToFolder(absoluteFilePath);
         profile.put("CoverPhoto", savedImagePath);
@@ -45,11 +34,21 @@ public class ProfileManager {
         ProfileUpdater.saveProfiles(profiles);
     }
     public void updateBio(String Bio){
-        JSONObject profile = validation.findProfileByUserId(userId);
+        JSONObject profile = findProfileByUserId(userId);
         JSONArray profiles = ProfileLoader.loadProfiles();
         profile.put("Bio", Bio);
         ProfileUpdater.updateProfile(userId, profiles, profile);
         ProfileUpdater.saveProfiles(profiles);
+    }
+    public JSONObject findProfileByUserId(String userId) {
+        JSONArray profiles = ProfileLoader.loadProfiles();
+        for (int i = 0; i < profiles.length(); i++) {
+            JSONObject profile = profiles.getJSONObject(i);
+            if (profile.getString("userId").equals(userId)) {
+                return profile;
+            }
+        }
+        return null;
     }
 
 }
